Extract Express routes
